Put DNS takeover in the enumerate stage
Takeover detection actively probes the third-party provider a record
dangles at, which is active information gathering about a discovered
asset rather than passive discovery — enumerate, not discover.

Restores the enumerate stage, which the zone transfer removal had
emptied. osintscan is now discover plus enumerate, with no pentest.
