+++ cmd/discover.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	// Standard
-	"errors"
 	"fmt"
 	"net"
 	"os"
@@ -22,7 +21,6 @@ import (
 	cctld "github.com/Method-Security/osintscan/internal/discover/dns/cctld"
 	subdomain "github.com/Method-Security/osintscan/internal/discover/dns/subdomain"
 	subdomainpassive "github.com/Method-Security/osintscan/internal/discover/dns/subdomain/passive"
-	takeover "github.com/Method-Security/osintscan/internal/discover/dns/takeover"
 	ip "github.com/Method-Security/osintscan/internal/discover/ip"
 
 	// External
@@ -95,7 +93,6 @@ func (a *OsintScan) InitDiscoverCommand() {
 	//   - active
 	//   - correlation
 	//   - passive
-	// - takeover
 	discoverDNSCmd := &cobra.Command{
 		Use:   "dns",
 		Short: "Gather intelligence on DNS services and assets",
@@ -662,99 +659,6 @@ func (a *OsintScan) InitDiscoverCommand() {
 	// Add command to 'dns' command
 	discoverDNSCmd.AddCommand(discoverDNSCctldCmd)
 
-	// Takeover Command
-	discoverDNSTakeoverCmd := &cobra.Command{
-		Use:   "takeover",
-		Short: "Detect potential subdomain takeovers",
-		Long: `Analyze the provided targets to identify DNS records that may be vulnerable to subdomain takeover attacks, using known fingerprints and heuristics.
-
-This detects a claimable record; it does not claim it. The CNAME lookup and the
-HTTP request both go to the third-party provider the record dangles at, never to
-infrastructure the target still controls.`,
-		Run: func(cmd *cobra.Command, args []string) {
-			// Get targets
-			targets, err := cmd.Flags().GetStringSlice("targets")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			filePaths, err := cmd.Flags().GetStringSlice("target-files")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			fileTargets, err := utils.GetEntriesFromFiles(filePaths)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			allTargets := append(targets, fileTargets...)
-			if len(allTargets) == 0 {
-				a.OutputSignal.AddError(errors.New("no targets specified"))
-				return
-			}
-
-			// Config
-			fingerprintsPath, err := cmd.Flags().GetString("fingerprints-file")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			fingerprints, err := takeover.RetrieveFingerprints(fingerprintsPath)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			if len(fingerprints) == 0 {
-				a.OutputSignal.AddError(errors.New("no fingerprints found"))
-				return
-			}
-
-			successfulOnly, err := cmd.Flags().GetBool("successful-only")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			verifyTLS, err := cmd.Flags().GetBool("verify-tls")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			timeout, err := cmd.Flags().GetInt("timeout")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			config := getDNSTakeoverConfig(allTargets, fileTargets, fingerprintsPath, successfulOnly, verifyTLS, timeout)
-
-			report, err := takeover.DetectDomainTakeover(cmd.Context(), config, fingerprints)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			a.OutputSignal.Content = report
-		},
-	}
-
-	// Target Flags
-	discoverDNSTakeoverCmd.Flags().StringSlice("targets", []string{}, "A list of URLs or domains to analyze for takeover vulnerabilities")
-	discoverDNSTakeoverCmd.Flags().StringSlice("target-files", []string{}, "File paths containing lists of targets to analyze for takeover vulnerabilities")
-	// Config Flags
-	discoverDNSTakeoverCmd.Flags().String("fingerprints-file", "", "Path to the JSON file containing service fingerprints for takeover detection")
-	discoverDNSTakeoverCmd.Flags().Bool("successful-only", false, "Show only confirmed successful takeovers in the results")
-	discoverDNSTakeoverCmd.Flags().Bool("verify-tls", false, "Verify TLS certificates when making HTTPS requests during takeover analysis")
-	discoverDNSTakeoverCmd.Flags().Int("timeout", 180, "Timeout in seconds for each takeover check request")
-
-	// Mark Required Flags
-	_ = discoverDNSTakeoverCmd.MarkFlagRequired("targets")
-
-	// Add command to 'dns' command
-	discoverDNSCmd.AddCommand(discoverDNSTakeoverCmd)
-
 	// Shodan Command
 	// Subcommands:
 	// - hostname
@@ -1170,24 +1074,4 @@ func getDiscoverDNSCctldConfig(domain string, cctlds []string, cctldsPreset stri
 	}
 
 	return config, nil
-}
-
-func getDNSTakeoverConfig(targets []string, targetFiles []string, fingerprintsFile string, successfulOnly bool, verifyTLS bool, timeout int) dnsfern.DiscoverDomainTakeoverConfig {
-	config := dnsfern.DiscoverDomainTakeoverConfig{
-		Targets:        targets,
-		TargetFiles:    targetFiles,
-		SuccessfulOnly: successfulOnly,
-		VerifyTls:      verifyTLS,
-	}
-	if fingerprintsFile != "" {
-		config.FingerprintsFile = &fingerprintsFile
-	}
-
-	if timeout >= 0 {
-		config.Timeout = timeout
-	} else {
-		config.Timeout = 30
-	}
-
-	return config
 }
