+++ main.go
@@ -15,6 +15,7 @@ func main() {
 	osintscan := cmd.NewOsintScan(version)
 	osintscan.InitRootCommand()
 	osintscan.InitDiscoverCommand()
+	osintscan.InitEnumerateCommand()
 
 	if err := osintscan.RootCmd.Execute(); err != nil {
 		os.Exit(1)
